+++ dave/server/debuggers/gdb/__init__.py
@@ -4,7 +4,7 @@ import dave.common.server_type as st
 st.SERVER_TYPE = st.ServerType.GDB
 
 from ...languages import *
-from .commands import GdbCommand, exit_handler, stop_handler, frame_checker
+from .commands import GdbCommand, exit_handler, stop_handler
 
 
 import gdb  # type: ignore
@@ -14,6 +14,5 @@ import os
 GdbCommand()
 gdb.events.exited.connect(exit_handler)
 gdb.events.stop.connect(stop_handler)
-gdb.events.before_prompt.connect(frame_checker)
 
-Logger().get().info("[dave] Successfully loaded")
+gdb.write("INFO : [dave] Successfully loaded\n")